polarplot: Fix QVector type from int to float
This is necessary since the input changed from mm to m

+++ src/waterfall/polarplot.h
@@ -116,7 +116,7 @@ private:
      */
     void updateMouseColumnData();
 
-    QVector<int> _distances;
+    QVector<float> _distances;
     QImage _image;
     float _maxDistance;
     float _mouseSampleAngle;
